TWINFACES-0 :rotating_light: Apply import/no-cycle rule
