+++ .eslintrc.json
@@ -7,6 +7,7 @@
   ],
   "plugins": ["fsd-import"],
   "rules": {
+    "import/no-cycle": ["error", { "maxDepth": 1 }],
     "fsd-import/layer-imports": [
       "error",
       {
