+++ src/entities/featurer/libs/constants.ts
@@ -3,13 +3,6 @@ import { FeaturerParamType } from "./types";
 
 export const ENTITY_COLOR = "#0EA5E9";
 
-export const FeaturerTypes = {
-  fieldTyper: 13,
-  trigger: 15,
-  validator: 16,
-  headHunter: 26,
-} as const;
-
 export const FEATURER_FIELD_SCHEMA = z.object({
   id: z.number().optional(),
   params: z
